[frontend] fix shop coupon
